recipes/llamaindex/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "llamaindex-cookbook"
version = "0.1.0"
description = ""
authors = ["Andrei Fajardo <andrei@llamaindex.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index-core = "^0.10.57"
llama-index-llms-openai = "^0.1.27"
llama-index-agent-openai = "^0.2.9"
streamlit = "^1.36.0"
llama-index-readers-file = "^0.1.30"
llama-index-embeddings-openai = "^0.1.11"
llama-index-program-openai = "^0.1.7"
snowflake-sqlalchemy = "^1.6.1"
llama-agents = {version = "^0.0.13", extras = ["rabbitmq"]}

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
pylint = "^3.2.5"
pytest = "^8.2.2"
black = {version = "^24.4.2", extras = ["jupyter"]}
codespell = {version = "^2.3.0", extras = ["toml"]}
isort = "^5.13.2"
ruff = "^0.5.1"
ipykernel = "^6.29.5"
